backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
tenacity==8.2.3
pydantic-settings==2.2.1
python-dotenv==1.0.1
asyncpg==0.29.0
email-validator==2.1.1
pyotp==2.9.0
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0
sqlalchemy-utils==0.41.1
bcrypt>=4.1.2,<5.0.0
greenlet==3.0.3
pandas==2.2.1
openpyxl==3.1.2
reportlab==4.1.0